@@ -1,25 +1,24 @@
 <?php
 
-use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
-     *
-     * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->id();
             $table->string('name');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->enum('active', ['Y', 'N']);
+            $table->enum('active', ['y', 'n'])->default('y');
+            $table->foreignId('theme_id')->constrained()->cascadeOnUpdate()->restrictOnDelete();
             $table->rememberToken();
             $table->timestamps();
         });
@@ -27,11 +26,13 @@ class CreateUsersTable extends Migration
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
-    public function down()
+    public function down(): void
     {
+        Schema::table('users', function (Blueprint $table) {
+            $table->dropForeign(['theme_id']);
+        });
+        
         Schema::dropIfExists('users');
     }
-}
+};
