@@ -1,5 +1,7 @@
 @extends('layouts.app')
 
+@section('title', __('Verify Your Email Address'))
+
 @section('content')
 <div class="container">
     <div class="row justify-content-center">
@@ -15,7 +17,11 @@
                     @endif
 
                     {{ __('Before proceeding, please check your email for a verification link.') }}
-                    {{ __('If you did not receive the email') }}, <a href="{{ route('verification.resend') }}">{{ __('click here to request another') }}</a>.
+                    {{ __('If you did not receive the email') }},
+                    <form class="d-inline" method="POST" action="{{ route('verification.resend') }}">
+                        @csrf
+                        <button type="submit" class="btn btn-link p-0 m-0 align-baseline">{{ __('click here to request another') }}</button>.
+                    </form>
                 </div>
             </div>
         </div>
